Middleware authentication using NextAuth

+++ src/actions/actions.ts
@@ -1,16 +1,41 @@
 "use server";
 import prisma from "@/lib/db";
 import { sleep } from "@/lib/utils";
-import { PetEssentials } from "@/lib/types";
 import { revalidatePath } from "next/cache";
-import { Pet } from "@prisma/client";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
+import { signIn, signOut } from "@/lib/auth";
 
-export async function addPet(pet: PetEssentials) {
+// -- User actions --
+
+//formadata from auth-form.tsx
+export async function logIn(formData: FormData) {
+  const authData = Object.fromEntries(formData.entries());
+
+  await signIn("credentials", authData);
+}
+
+export async function logOut() {
+  await signOut({
+    redirectTo: "/",
+  });
+}
+
+// -- Pet actions --
+
+export async function addPet(pet: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+
+  if (!validatedPet.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
 
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -21,15 +46,24 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(newPetData);
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPet.success || !validatedPetId.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
 
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -40,13 +74,21 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: "Invalid pet data",
+    };
+  }
 
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
